fix(cache): set expiry on updates
The cache may have entries that have yet to expire.

If we attempt to insert a value with a key that matches any existing
entries, the default implementation keeps the original cache expiration
instead of updating it.

By overriding `expire_after_update`, we ensure that expiration is
properly updated when updating cache values for existing keys.