refactor: use json-subscriber for improved json logs
Includes of Trace and Span IDs from OpenTelemetry.
This allows us to exclude the current and all involved spans from the
trace event.

The crate also allows for flattening the spans for
deduplication, should we want to also log the span attributes.
Seeing as we use OpenTelemetry, it should suffice with just the IDs for now.